added time to read calculation

+++ src/Models/Post.cs
@@ -17,7 +17,7 @@ namespace Deferat.Models
         public bool Featured { get; set; }
         public bool Draft { get; set; }
         public string ShortContent { get; set; }
-        public long TimeToRead { get; set; }
+        public double TimeToRead { get; set; }
     }
 
 }
